+++ src/App.jsx
@@ -5,9 +5,11 @@ import PhotoReels from "./pages/photoreel";
 import Header from "./components/header";
 
 import Footer from "./components/footer";
-import PortfolioPage from "./pages/portfolio";
+import PortfolioPage from "./pages/portfolio/portfolio";
 import IndividualPortfolio from "./pages/portfolio/individualPortfolio";
 import { ThemeProvider } from "./context/ThemeContext";
+import PortfolioLayout from "./pages/portfolio/index";
+
 
 export default function App() {
   return (
@@ -17,9 +19,11 @@ export default function App() {
         <div className="pt-[65px]">
           <Routes>
             <Route path="/" element={<HomePage />} />
-            <Route path="/portfolio" element={<PortfolioPage />} />
-            <Route path="/portfolio/:id" element={<IndividualPortfolio />} />
             <Route path="/photo-reels" element={<PhotoReels />} />
+            <Route path="/portfolio" element={<PortfolioLayout />}>
+              <Route index element={<PortfolioPage />} />
+              <Route path=":id" element={<IndividualPortfolio />} />
+            </Route>
           </Routes>
         </div>
         <Footer />
